backend/requirements.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.7  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
flower==0.9.5  # https://github.com/mher/flower
uvicorn==0.11.8  # https://github.com/encode/uvicorn
psutil==5.7.2
ipip-ipdb==1.3.2
itsdangerous==1.1.0

# Django
# ------------------------------------------------------------------------------
Django==3.1.7  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-extensions==3.0.5
django-filter==2.4.0

# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
#django-cors-headers==3.4.0 # https://github.com/adamchainz/django-cors-headers

# Prod
gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
mysqlclient==2.0.3
Collectfast==2.2.0  # https://github.com/antonagestam/collectfast